Move columns to app
