+++ frontend/scripts/react-components/tool-links/tool-links.actions.js
@@ -1,8 +1,6 @@
 import castArray from 'lodash/castArray';
 
 export const TOOL_LINKS__SET_FLOWS_LOADING = 'TOOL_LINKS__SET_FLOWS_LOADING';
-export const TOOL_LINKS__GET_COLUMNS = 'TOOL_LINKS__GET_COLUMNS';
-export const TOOL_LINKS__SET_COLUMNS = 'TOOL_LINKS__SET_COLUMNS';
 export const TOOL_LINKS__SET_NODES = 'TOOL_LINKS__SET_NODES';
 export const TOOL_LINKS__SET_LINKS = 'TOOL_LINKS__SET_LINKS';
 export const TOOL_LINKS__SELECT_VIEW = 'TOOL_LINKS__SELECT_VIEW';
@@ -25,19 +23,6 @@ export function setToolFlowsLoading(loading) {
   return {
     type: TOOL_LINKS__SET_FLOWS_LOADING,
     payload: { loading }
-  };
-}
-
-export function getToolColumns() {
-  return {
-    type: TOOL_LINKS__GET_COLUMNS
-  };
-}
-
-export function setToolColumns(columns) {
-  return {
-    type: TOOL_LINKS__SET_COLUMNS,
-    payload: { columns }
   };
 }
