+++ frontend/scripts/react-components/tool-links/tool-links.initial-state.js
@@ -2,7 +2,6 @@
 
 export default {
   data: {
-    columns: null,
     nodes: null,
     links: null,
     nodeHeights: null,
