+++ frontend/scripts/react-components/tool-links/tool-links.saga.js
@@ -1,5 +1,5 @@
 import { select, all, call, fork, put, takeLatest, cancel } from 'redux-saga/effects';
-import { SET_CONTEXT, SET_CONTEXTS } from 'actions/app.actions';
+import { SET_CONTEXT, SET_CONTEXTS, TOOL_LINKS__GET_COLUMNS } from 'actions/app.actions';
 import { setLoadingSpinner } from 'utils/saga-utils';
 import { loadMapVectorData, SELECT_YEARS } from 'react-components/tool/tool.actions';
 import { getSelectedContext } from 'reducers/app.selectors';
@@ -7,7 +7,6 @@ import {
   TOOL_LINKS__SET_SELECTED_NODES,
   TOOL_LINKS__SELECT_COLUMN,
   TOOL_LINKS__SELECT_VIEW,
-  TOOL_LINKS__GET_COLUMNS,
   TOOL_LINKS__EXPAND_SANKEY,
   TOOL_LINKS__COLLAPSE_SANKEY,
   TOOL_LINKS__CLEAR_SANKEY,
