+++ frontend/scripts/react-components/tool/columns-selector-group/columns-selector-group.container.js
@@ -3,7 +3,7 @@ import ColumnsSelectorGroup from 'react-components/tool/columns-selector-group/c
 
 const mapStateToProps = state => ({
   sankeySize: state.toolLayers.sankeySize,
-  columns: state.toolLinks.data.columns
+  columns: state.app.data.columns
 });
 
 export default connect(mapStateToProps)(ColumnsSelectorGroup);
