+++ frontend/scripts/reducers/app.initial-state.js
@@ -6,6 +6,9 @@ export default {
   tooltips: null,
   contextIsUserSelected: !SHOW_WORLD_MAP_IN_EXPLORE,
   currentDropdown: null,
+  data: {
+    columns: null
+  },
   modal: {
     visibility: false,
     modalParams: null
